+++ CHANGELOG.md
@@ -37,6 +37,10 @@
 - (!) Replaced the template keyword "args" with "settings" in yaml template files.
 - Changed the default value of the "evaluate_log2_transformation" setting to False.
 - Changed the default value of the "write_supheader" setting to False.
+- The "sample_extraction_tag" setting is now used as a regular expression pattern
+  instead of looking for an exact match.
+- The "tag" parameter of the "tag sample section" is now used as a regular expression
+  pattern instead of simply looking for an exact match.
 - Supheader cells are now merged even when no supheader name is specified.
 - When attempting to use an invalid format for compiling an Excel report, the compiler
   will now apply no format and issue a warning instead of raising an exception. This